Image_classification/fruit_classification/requirements.txt
flask==2.0.3
torch==1.11.0
torchvision==0.12.0+cpu
torchaudio==0.11.0
numpy==1.22.0
pillow==9.0.1
